github_access/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic-settings==2.3.4
PyGithub==2.2.0
google-generativeai==0.7.0
python-dotenv==1.0.1
tree-sitter==0.21.1
tree-sitter-languages==1.10.0
python-radon==6.0.0
tenacity==8.4.2
python-multipart==0.0.9
